Split up write_image.cpp and remove redundant use of InsetState
